Fix the Scanner Bug
The Link Address Menu option called the linkAddress method in the Customer class where a new scanner object  was being opened on System.in stream. This was causing the scanner object in the main method to malfunction when the control was being passed from the linkAddress method in the Customer class, back to the main method in the Tester class.

The bug was causing a NoSuchElementFound Exception on the scanner object in the main method.

I fixed the bug by passing the same scanner object which was open in the main method to the linkAddress method in the Customer class. This resolved the conflict and the Scanner object was behaving properly again.

Thanks to Tanmay, Aditya, Sumit,
 and Pavan for helping me fix the bug.